--- a/Cargo.toml
+++ b/Cargo.toml
@@ -14,7 +14,6 @@ travis-ci = { repository = "antifuchs/ratelimit_meter", branch = "master" }
 maintenance = { status = "passively-maintained" }
 
 [dependencies]
-crossbeam = "0.3.0"
 failure = "0.1.1"
 failure_derive = "0.1.1"